+++ lib/ui/BrowsePanel.dart
@@ -502,7 +502,7 @@ class _BrowseEntry extends StatelessWidget {
       case "my.my_laundry":      _onTapMyLaundry(context); break;
       case "my.my_inbox":        _onTapMyNotifications(context); break;
       case "my.my_campus_guide": _onTapMyCampusGuide(context); break;
-      case "my.my_wellness_resources": _onTapMyWellnessResources(context); break;
+      case "my.wellness_resources": _onTapWellnessResources(context); break;
 
       case "polls.create_poll":  _onTapCreatePoll(context); break;
       case "polls.recent_polls":   _onTapViewPolls(context); break;
@@ -522,7 +522,7 @@ class _BrowseEntry extends StatelessWidget {
       case "wellness.rings": _onTapWellnessRings(context); break;
       case "wellness.todo":  _onTapWellnessToDo(context); break;
       case "wellness.tips":  _onTapWellnessTips(context); break;
-      case "wellness.my_wellness_resources": _onTapMyWellnessResources(context); break;
+      case "wellness.wellness_resources": _onTapWellnessResources(context); break;
     }
   }
 
@@ -818,8 +818,8 @@ class _BrowseEntry extends StatelessWidget {
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return SavedPanel(favoriteCategories: [GuideFavorite.favoriteKeyName]); } ));
   }
 
-  void _onTapMyWellnessResources(BuildContext context) {
-    Analytics().logSelect(target: "My Wellness Resources");
+  void _onTapWellnessResources(BuildContext context) {
+    Analytics().logSelect(target: "Wellness Resources");
     Navigator.push(context, CupertinoPageRoute(builder: (context) { return WellnessHomePanel(content: WellnessContent.resources,); } ));
   }